requirements.txt
requests>=2.28
jieba>=0.42
pypinyin>=0.49
pycantonese>=3.4
pyyaml>=6.0
